tests/wpt/web-platform-tests/css/css-ui/resize-015.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Basic User Interface Test: resizing uses the style attribute</title>
<link rel="author" title="Florian Rivoal" href="http://florian.rivoal.net">
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#resize">
<meta name="flags" content="interact">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#test {
  position: absolute;
  width: 100px;
  height: 100px;
  background: orange;
  overflow: hidden;
  resize: both;
}
#target {
  position: absolute;
  width: 150px;
  height: 150px;
  background: blue;
}
#log { margin-top: 200px; }
</style>

<p>Enlarge the orange box so that it at least covers the blue box entirely, then press the “done” button.</p>
<!-- There's no actual need to cover the whole box,
  but doing so makes sure that the user has resized in both dimensions -->

<button onclick="verify()">done</button>

<div id="target"></div>
<div id="test"></div>
<div id=log></div>
<script>
setup({ explicit_done: true });
function verify() {
  test(
    function(){
      var test = document.getElementById("test");
      assert_regexp_match(test.style.width, /px$/, "The width property of the style attribute should be set in pixels");
      assert_not_equals(test.style.width, "100px", "The width should be different from the initial one");
      assert_regexp_match(test.style.height, /px$/, "The height property of the style attribute should be set in pixels");
      assert_not_equals(test.style.height, "100px", "The height should be different from the initial one");
    }, "resize works by setting the width and height properties on the style attrbute in pixels");
  test(
    function(){
      var test = document.getElementById("test");
      assert_false(test.getAttribute("style").includes("important"));
    }, "resize does not set !important");
  done();
}
</script>
